refactor: move shared components into dc-ui
